Modify original OfflineAnalsis.Rmd file
Change model names and dataset names so that they were different for
easier comparison. Created separate data frames for each dataset being
modelled so that the dataset could be called as a parameter in the goodness
of fit test. Add in code for Hosmer-Lemeshow goodness of
fit test at line 98.